refactor: moved utility functions to a utils file
