+++ packages/react/src/components/Authenticator/FederatedIdentities/controls/FederatedIdentitiesControl.tsx
@@ -4,14 +4,11 @@ import {
   CreateFederatedIdentitiesInput,
   RenderButton,
   UseHandleSignInWithRedirect,
-  ProviderType,
-  ProviderData,
-  DefaultFederatedProviderList,
   CreateProviderInput,
 } from './types';
 import { FederatedIdentitiesElements } from '../context';
 import createProvider from './createProvider';
-import { capitalize, FederatedProvider } from '@aws-amplify/ui';
+import { toProviderData } from './utils';
 
 interface ChildrenProps {
   children?: React.ReactNode;
@@ -32,44 +29,6 @@ type FederatedIdentitiesProps<T extends string = string> = (
 export interface FederatedIdentities<T extends string = string>
   extends ForwardRefExoticComponent<FederatedIdentitiesProps<T>> {
   Identity: IdentityControl<T>;
-}
-
-function getSupportedProviderData<T extends string = string>(
-  providerName: T
-): ProviderData<T> {
-  return {
-    displayName: capitalize<T>(providerName),
-    icon: providerName,
-    providerName: providerName,
-  };
-}
-
-function validateProviderTypes(providers: ProviderType[]): void {
-  const providerNames = new Set<string>();
-
-  providers.forEach((provider) => {
-    const providerName =
-      typeof provider === 'string' ? provider : provider.providerName;
-
-    if (providerNames.has(providerName)) {
-      throw new Error(`Duplicate provider name found: ${providerName}`);
-    }
-
-    providerNames.add(providerName);
-  });
-}
-
-export function toProviderData<T extends string = string>(
-  providers: ProviderType<T>[]
-): ProviderData<T>[] {
-  validateProviderTypes(providers);
-  return providers.map((provider) => {
-    if (DefaultFederatedProviderList.includes(provider as FederatedProvider)) {
-      return getSupportedProviderData<T>(provider as T);
-    } else {
-      return provider as ProviderData<T>;
-    }
-  });
 }
 
 export function createFederatedIdentities<
